fixed typo, and rename component
