+++ src/components/LandingPage/index.js
@@ -10,8 +10,8 @@ import ExperienceBlock from '../ExperienceSearch/ExperienceBlock';
 import { fetchExperiences } from '../../actions/experienceSearch';
 import { fetchMetaListIfNeeded } from '../../actions/laborRightsMenu';
 import LaborRightsEntry from '../LaborRightsMenu/LaborRightsEntry';
-import HomeBanner from './HomeBanner';
-import HomeData from './HomeData';
+import Banner from './Banner';
+import Dashboard from './Dashboard';
 import helmetData from '../../constants/helmetData';
 
 class LandingPage extends Component {
@@ -48,8 +48,8 @@ class LandingPage extends Component {
     return (
       <main>
         {this.renderHelmet()}
-        <HomeBanner />
-        <HomeData />
+        <Banner />
+        <Dashboard />
         <Section padding>
           <Wrapper size="l">
             <Heading size="l" center marginBottom>熱門分享</Heading>
