+++ src/components/Layout/App.module.css
@@ -19,7 +19,7 @@
 /* gloabl style start */
 body, html {
 	font-family: primary-font;
-	font-weight: 00;
+	font-weight: 400;
 	color: #444;
 	font-size: 16px;
 	-webkit-tap-highlight-color: rgba(0,0,0,.2);
